_display subpackage
