+++ limix/qtl/_iscan.py
@@ -1,10 +1,10 @@
 from numpy import concatenate, newaxis
 from tqdm import tqdm
 
-from limix.display import timer_text
+from limix._display import timer_text
 
 from .._data import conform_dataset
-from ..display import session_text
+from .._display import session_text
 
 from glimix_core.lmm import LMM
 from .._likelihood import assert_likelihood_name, normalise_extreme_values
