+++ limix/qtl/_scan.py
@@ -2,10 +2,10 @@ from __future__ import division
 
 import sys
 
-from limix.display import timer_text
+from limix._display import timer_text
 
 from .._data import conform_dataset
-from ..display import session_text
+from .._display import session_text
 from .._likelihood import assert_likelihood_name, normalise_extreme_values
 from ._model import QTLModel
